+++ app/src/data/encoding.js
